requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.7
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.4.0
